[refactor] Change folderId type to Long

+++ data/src/main/java/daily/dayo/data/datasource/remote/post/PostRequest.kt
@@ -9,7 +9,7 @@ data class EditPostRequest (
     @SerializedName("contents")
     val contents : String,
     @SerializedName("folderId")
-    val folderId : Int,
+    val folderId : Long,
     @SerializedName("hashtags")
     val hashtags : List<String>
 )
